Refactor extension activation auth gate
