+++ crates/ironclaw_reborn_composition/src/factory.rs
@@ -891,8 +891,9 @@ async fn build_local_dev(input: RebornBuildInput) -> Result<RebornServices, Rebo
         }
     };
     services = services.with_runtime_credential_account_resolver(Arc::new(
-        ProductAuthRuntimeCredentialResolver::new(
+        ProductAuthRuntimeCredentialResolver::new_with_refresh(
             product_auth.runtime_credential_account_selection_service(),
+            product_auth.runtime_credential_account_refresh_service(),
         ),
     ));
     let mut available_extensions = AvailableExtensionCatalog::from_filesystem_root(
@@ -2914,8 +2915,9 @@ where
     // CredentialAccountService. Unconditional in production: product_auth_services
     // always exists (durable filesystem fallback from #4234).
     let services = services.with_runtime_credential_account_resolver(Arc::new(
-        ProductAuthRuntimeCredentialResolver::new(
+        ProductAuthRuntimeCredentialResolver::new_with_refresh(
             product_auth_services.runtime_credential_account_selection_service(),
+            product_auth_services.runtime_credential_account_refresh_service(),
         ),
     ));
     register_bundled_gsuite_first_party_handlers(
@@ -3596,7 +3598,7 @@ mod tests {
             .await
             .expect("install Gmail");
         extension_management
-            .activate_without_credential_preflight_for_test(
+            .activate_with_prechecked_credentials_for_test(
                 gmail_ref,
                 ExtensionActivationMode::Static,
             )
@@ -3607,7 +3609,7 @@ mod tests {
             .await
             .expect("install Google Calendar");
         extension_management
-            .activate_without_credential_preflight_for_test(
+            .activate_with_prechecked_credentials_for_test(
                 calendar_ref,
                 ExtensionActivationMode::Static,
             )
@@ -3731,7 +3733,7 @@ mod tests {
             .await
             .expect("install Notion MCP");
         extension_management
-            .activate_without_credential_preflight_for_test(
+            .activate_with_prechecked_credentials_for_test(
                 notion_ref,
                 ExtensionActivationMode::Static,
             )
@@ -3785,7 +3787,7 @@ mod tests {
             .await
             .expect("install Web Access");
         extension_management
-            .activate_without_credential_preflight_for_test(
+            .activate_with_prechecked_credentials_for_test(
                 web_access_ref,
                 ExtensionActivationMode::Static,
             )
