+++ crates/ironclaw_reborn_composition/src/lib.rs
@@ -33,6 +33,8 @@ mod budget_events;
 mod bundled_skills;
 mod default_system_prompt;
 mod error;
+mod extension_activation_credentials;
+mod extension_credential_requirements;
 mod extension_installation_store;
 mod extension_lifecycle;
 mod extension_lifecycle_capabilities;
